+++ accumulo-spark/src/main/scala/geotrellis/spark/store/accumulo/AccumuloLayerReindexer.scala
@@ -21,7 +21,6 @@ import geotrellis.store._
 import geotrellis.store.accumulo._
 import geotrellis.store.avro._
 import geotrellis.store.index._
-import geotrellis.layer.json._
 import geotrellis.spark.store._
 import geotrellis.util._
