+++ accumulo/src/main/scala/geotrellis/store/accumulo/AccumuloCollectionLayerProvider.scala
@@ -21,10 +21,7 @@ import geotrellis.store._
 import geotrellis.store.accumulo.conf.AccumuloConfig
 import geotrellis.util.UriUtils
 
-import org.apache.spark.SparkContext
-
 import java.net.URI
-
 
 /**
  * Provides [[AccumuloAttributeStore]] instance for URI with `accumulo` scheme.
