+++ s3-spark/src/main/scala/geotrellis/spark/store/s3/S3LayerManager.scala
@@ -21,7 +21,6 @@ import geotrellis.store._
 import geotrellis.store.AttributeStore.Fields
 import geotrellis.store.avro._
 import geotrellis.store.index._
-import geotrellis.layer.json._
 import geotrellis.store.s3._
 import geotrellis.util._
 import geotrellis.spark._
@@ -29,7 +28,7 @@ import geotrellis.spark.store._
 
 import software.amazon.awssdk.services.s3.S3Client
 import org.apache.spark.SparkContext
-import io.circe._
+import _root_.io.circe._
 
 import scala.reflect.ClassTag
