+++ s3-spark/src/test/scala/geotrellis/spark/store/s3/S3SpatialSpec.scala
@@ -58,7 +58,7 @@ class S3SpatialSpec
   lazy val writer = new S3LayerWriter(attributeStore, bucket, prefix, identity, getS3Client)
   lazy val deleter = new S3LayerDeleter(attributeStore, getS3Client)
   lazy val copier  = new S3LayerCopier(attributeStore, bucket, prefix, getS3Client)
-  lazy val reindexer = GenericLayerReindexer[S3LayerHeader](attributeStore, reader, writer, deleter, copier)
+  lazy val reindexer = GenericLayerReindexer(attributeStore, reader, writer, deleter, copier)
   lazy val mover = GenericLayerMover(copier, deleter)
   lazy val tiles = new S3ValueReader(attributeStore, getS3Client)
   lazy val sample = AllOnesTestFile
