--- s3-testkit/src/main/scala/geotrellis/spark/io/s3/testkit/MockS3LayerWriter.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2016 Azavea
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.spark.io.s3.testkit
-
-import geotrellis.spark.io._
-import geotrellis.spark.io.s3._
-
-class MockS3LayerWriter(
-  attributeStore: AttributeStore,
-  bucket: String,
-  keyPrefix: String
-) extends S3LayerWriter(attributeStore, bucket, keyPrefix) {
-  override def rddWriter =
-    new S3RDDWriter {
-      def getS3Client = () => {
-        new MockS3Client()
-      }
-    }
-}
