+++ s3/src/main/scala/geotrellis/store/s3/S3LayerMover.scala
@@ -21,7 +21,7 @@ import geotrellis.store._
 import geotrellis.util._
 
 import software.amazon.awssdk.services.s3.S3Client
-k
+
 import scala.reflect.ClassTag
 
 object S3LayerMover {
