+++ s3/src/main/scala/geotrellis/store/s3/S3ValueReader.scala
@@ -28,7 +28,7 @@ import software.amazon.awssdk.services.s3.model.{S3Exception, GetObjectRequest}
 import software.amazon.awssdk.services.s3.S3Client
 import org.apache.avro.Schema
 import org.apache.commons.io.IOUtils
-import io.circe._
+import _root_.io.circe._
 
 import scala.reflect.ClassTag
