+++ spark/src/test/scala/geotrellis/spark/store/CoordinateSpaceTimeSpec.scala
@@ -20,7 +20,6 @@ import geotrellis.layer._
 import geotrellis.raster.{GridBounds, Tile}
 import geotrellis.store._
 import geotrellis.store.avro.codecs._
-import geotrellis.layer.json._
 import geotrellis.spark._
 import geotrellis.spark.testkit.io._
 import geotrellis.spark.testkit.testfiles.TestFiles
