+++ spark/src/test/scala/geotrellis/spark/store/CoordinateSpaceTimeTileFeatureSpec.scala
@@ -20,12 +20,10 @@ import geotrellis.layer._
 import geotrellis.raster.{GridBounds, Tile, TileFeature}
 import geotrellis.store._
 import geotrellis.store.avro.codecs._
-import geotrellis.layer.json._
 import geotrellis.spark._
 import geotrellis.spark.testkit.io._
 
 import java.time.{ZoneOffset, ZonedDateTime}
-
 
 trait CoordinateSpaceTimeTileFeatureSpec { self: PersistenceSpec[SpaceTimeKey, TileFeature[Tile, Tile], TileLayerMetadata[SpaceTimeKey]] =>
   val dates = Vector( // all the dates in the layer
